update linking module
