+++ packages/expo-linking/src/ExpoLinking.web.ts
@@ -1,23 +1,17 @@
-import { Platform } from 'expo-modules-core';
 import invariant from 'invariant';
 
 import { NativeURLListener, URLListener } from './Linking.types';
-
-const EventTypes = ['url'];
 
 const listeners: { listener: URLListener; nativeListener: NativeURLListener }[] = [];
 
 export default {
   addEventListener(type: 'url', listener: URLListener): { remove(): void } {
     // Do nothing in Node.js environments
-    if (!Platform.isDOMAvailable) {
+    if (typeof window === 'undefined') {
       return { remove() {} };
     }
 
-    invariant(
-      EventTypes.indexOf(type) !== -1,
-      `Linking.addEventListener(): ${type} is not a valid event`
-    );
+    invariant(type === 'url', `Linking.addEventListener(): ${type} is not a valid event`);
     const nativeListener: NativeURLListener = (nativeEvent) =>
       listener({ url: window.location.href, nativeEvent });
     listeners.push({ listener, nativeListener });
@@ -31,13 +25,10 @@ export default {
 
   removeEventListener(type: 'url', listener: URLListener): void {
     // Do nothing in Node.js environments
-    if (!Platform.isDOMAvailable) {
+    if (typeof window === 'undefined') {
       return;
     }
-    invariant(
-      EventTypes.indexOf(type) !== -1,
-      `Linking.removeEventListener(): ${type} is not a valid event.`
-    );
+    invariant(type === 'url', `Linking.addEventListener(): ${type} is not a valid event`);
     const listenerIndex = listeners.findIndex((pair) => pair.listener === listener);
     invariant(
       listenerIndex !== -1,
@@ -48,18 +39,18 @@ export default {
     listeners.splice(listenerIndex, 1);
   },
 
-  async canOpenURL(url: string): Promise<boolean> {
+  async canOpenURL(): Promise<boolean> {
     // In reality this should be able to return false for links like `chrome://` on chrome.
     return true;
   },
 
   async getInitialURL(): Promise<string> {
-    if (!Platform.isDOMAvailable) return '';
+    if (typeof window === 'undefined') return '';
     return window.location.href;
   },
 
   async openURL(url: string): Promise<void> {
-    if (Platform.isDOMAvailable) {
+    if (typeof window !== 'undefined') {
       // @ts-ignore
       window.location = new URL(url, window.location).toString();
     }
