+++ packages/expo-linking/src/Schemes.web.ts
@@ -1,0 +1,12 @@
+export function hasCustomScheme(): boolean {
+  return true;
+}
+export function collectManifestSchemes(): string[] {
+  return [];
+}
+export function hasConstantsManifest(): boolean {
+  return false;
+}
+export function resolveScheme(): string {
+  return 'https';
+}
